Move monitoring to .utils, change method to create_from_request
