+++ sentinelhub/api/batch/__init__.py
@@ -2,12 +2,6 @@
 The part of the package that implements interface with Sentinel Hub services.
 """
 from .base import BatchRequestStatus, BatchUserAction
-from .process import (
-    BatchCollection,
-    BatchRequest,
-    BatchTileStatus,
-    SentinelHubBatch,
-    monitor_batch_analysis,
-    monitor_batch_job,
-)
+from .process import BatchCollection, BatchRequest, BatchTileStatus, SentinelHubBatch
 from .statistical import BatchStatisticalRequest, SentinelHubBatchStatistical
+from .utils import monitor_batch_analysis, monitor_batch_job
